+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Switch from state log to changelog and improve notifications about changes and
+  updater crashes
+
+
 ## [1.4.3] - 2020-12-04
 ### Fixed
 - `packages` module failure when there is uninstalled package in index